First Chess Server Commit
So right now, this just starts two algorithms and has
them talk to each other. No chess logic is in place yet.

If you want to test it, set CHESS_ROOT to the ..../chess and the
make file will put play_game.out in CHESS_ROOT/bin.

a.out is the test algorithm.