Logic for Piece Moves
Added logic for determining whether a move is valid for a particular piece.

Fixed up BoardPosition to be relative to the bottom left of a chess board with
white on the bottom.

Added a check to ensure that a move that stays in the same place is invalid.

There was two definitions of White and Black, which was also fixed.